DRT-Sprint - cleanup code - restore default email senders
