+++ Equilobe.DailyReport.Models/Jira/Sprint.cs
@@ -1,29 +1,30 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace Equilobe.DailyReport.Models.Jira
 {
+    public class JiraSprints
+    {
+        public long maxResults { get; set; }
+        public long startAt { get; set; }
+        public long total { get; set; }
+        public bool isLast { get; set; }
+        public List<Sprint> values { get; set; }
+    }
+
     public class Sprint
     {
-        [DataMember]
-        public int id { get; set; }
-        [DataMember]
-        public int sequence { get; set; }
-        [DataMember]
-        public string name { get; set; }
-        [DataMember]
+        public string id { get; set; }
+        public string self { get; set; }
         public string state { get; set; }
-        [DataMember]
-        public int linkedPagesCount { get; set; }
-        [DataMember]
+        public string name { get; set; }
+        public string originalBoardId { get; set; }
+        public string goal { get; set; }
         public string startDate { get; set; }
-        [DataMember]
         public string endDate { get; set; }
-        [DataMember]
         public string completeDate { get; set; }
 
         public DateTime? StartDate
@@ -34,12 +35,13 @@ namespace Equilobe.DailyReport.Models.Jira
                 {
                     return Convert.ToDateTime(startDate);
                 }
-                catch(Exception)
+                catch (Exception)
                 {
                     return null;
                 }
             }
         }
+
         public DateTime? EndDate
         {
             get
@@ -63,7 +65,7 @@ namespace Equilobe.DailyReport.Models.Jira
                 {
                     return Convert.ToDateTime(completeDate);
                 }
-                catch(Exception)
+                catch (Exception)
                 {
                     return null;
                 }
