--- Equilobe.DailyReport.Models/Jira/Sprints.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.Serialization;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Equilobe.DailyReport.Models.Jira
-{
-    [DataContract]
-    public class Sprints
-    {
-        [DataMember]
-        public List<Sprint> sprints { get; set; }
-        [DataMember]
-        public string rapidViewId { get; set; }
-    }
-}
